refactor(perps): delegate getDisplaySymbol to controller's getPerpsDisplaySymbol
ui/components/app/perps/utils.ts duplicated the same HIP-3 prefix-stripping
logic already exported as getPerpsDisplaySymbol from @metamask/perps-controller.
Delegate to it instead of maintaining a parallel copy that could drift, while
keeping the getDisplaySymbol name so its ~20 existing call sites are unchanged.

Also adds getPerpsDisplaySymbol to the shared Jest mock for
@metamask/perps-controller, which didn't have it yet.
